+++ src/entity/user.rs
@@ -11,7 +11,7 @@ pub struct Model {
     pub email: Option<String>,
     pub first_name: Option<String>,
     pub last_name: Option<String>,
-    pub age: Option<i8>,
+    pub age: Option<i32>,
     pub country_id: Option<i32>,
     pub created_at: Option<DateTime>,
 }
